Parse logic and Server.java moved to views
